# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Restaurant Reservation API Documentation",
      "version": "1.0.0",
      "description": "API for managing restaurant table reservations, built with Node.js, Express, and MongoDB. Includes Google Authentication for secure access."
    },
    "servers": [
      {
        "url": "http://localhost:3000",
        "description": "Local development server"
      }
      ],
...
    "components": {
      "securitySchemes": {
        "cookieAuth": {
          "type": "apiKey",
          "in": "cookie",
          "name": "connect.sid",
          "description": "Session cookie for Google Authentication. After successful Google login, a `connect.sid` cookie is set, which is used for subsequent authenticated requests."
        }
      },
      "schemas": {
        "ErrorResponse": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "description": "Error message describing the issue."
            }
          },
          "example": {
            "message": "Validation error: Customer name is required."
          }
        },
        "Table": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "description": "Unique ID of the table (MongoDB ObjectId)",
              "readOnly": true
            },
            "tableNumber": {
              "type": "string",
              "description": "Unique identifier for the table (e.g., 'T1', '102')",
              "example": "T1"
            },
            "capacity": {
              "type": "integer",
              "format": "int32",
              "description": "Maximum capacity of the table",
              "example": 4
            },
            "location": {
              "type": "string",
              "description": "Location of the table within the restaurant (optional)",
              "example": "Indoor, near window"
            },
            "isAvailable": {
              "type": "boolean",
              "description": "Indicates if the table is currently available for reservations (optional, default: true)",
              "example": true
            }
          },
          "required": [
            "tableNumber",
            "capacity"
          ],
          "example": {
            "tableNumber": "T1",
            "capacity": 4,
            "location": "Indoor",
            "isAvailable": true
          }
        },
...
      }
    },
    "paths": {
      "/tables": {
        "post": {
          "tags": ["Tables"],
          "summary": "Create a new table",
          "security": [{"cookieAuth": []}],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Table"
                },
                "example": {
                  "tableNumber": "T10",
                  "capacity": 2,
                  "location": "Outdoor Patio"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Table created successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Table"
                  },
                  "example": {
                    "_id": "64d8c2e4f7b3e2a0c5d4e6f9",
                    "tableNumber": "T10",
                    "capacity": 2,
                    "location": "Outdoor Patio",
                    "isAvailable": true
                  }
                }
              }
            },
            "400": {
              "description": "Bad request - validation errors",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Table number is required."
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized - authentication required",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Authentication required to access this endpoint."
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Internal server error."
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": ["Tables"],
          "summary": "Get all tables",
          "security": [{"cookieAuth": []}],
          "responses": {
            "200": {
              "description": "Successful retrieval of tables",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Table"
                    }
                  },
                  "example": [
                    {
                      "_id": "64d8c2e4f7b3e2a0c5d4e6f8",
                      "tableNumber": "T1",
                      "capacity": 4,
                      "location": "Indoor",
                      "isAvailable": true
                    },
                    {
                      "_id": "64d8c2e4f7b3e2a0c5d4e6f9",
                      "tableNumber": "T2",
                      "capacity": 2,
                      "location": "Outdoor Patio",
                      "isAvailable": false
                    }
                  ]
                }
              }
            },
            "401": {
              "description": "Unauthorized - authentication required",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Authentication required to access this endpoint."
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Internal server error."
                  }
                }
              }
            }
          }
        }
      },
      "/tables/{tableId}": {
        "parameters": [
          {
            "name": "tableId",
            "in": "path",
            "required": true,
            "description": "ID of the table to perform operations on",
            "schema": {
              "type": "string"
            }
          }
        ],
        "get": {
          "tags": ["Tables"],
          "summary": "Get a specific table by ID",
          "security": [{"cookieAuth": []}],
          "responses": {
            "200": {
              "description": "Successful retrieval of table",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Table"
                  },
                  "example": {
                    "_id": "64d8c2e4f7b3e2a0c5d4e6f8",
                    "tableNumber": "T1",
                    "capacity": 4,
                    "location": "Indoor",
                    "isAvailable": true
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized - authentication required",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Authentication required to access this endpoint."
                  }
                }
              }
            },
            "404": {
              "description": "Table not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Table not found."
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Internal server error."
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": ["Tables"],
          "summary": "Update an existing table by ID",
          "security": [{"cookieAuth": []}],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Table"
                },
                "example": {
                  "tableNumber": "T1-Updated",
                  "capacity": 4,
                  "location": "Indoor, near window",
                  "isAvailable": false
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Table updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Table"
                  },
                  "example": {
                    "_id": "64d8c2e4f7b3e2a0c5d4e6f8",
                    "tableNumber": "T1-Updated",
                    "capacity": 4,
                    "location": "Indoor, near window",
                    "isAvailable": false
                  }
                }
              }
            },
            "400": {
              "description": "Bad request - validation errors",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Capacity must be a number."
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized - authentication required",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Authentication required to access this endpoint."
                  }
                }
              }
            },
            "404": {
              "description": "Table not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Table not found."
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Internal server error."
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["Tables"],
          "summary": "Delete a table by ID",
          "security": [{"cookieAuth": []}],
          "responses": {
            "204": {
              "description": "Table deleted successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Authentication required to access this endpoint."
                  }
                }
              }
            },
            "404": {
              "description": "Table not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Table not found."
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  },
                  "example": {
                    "message": "Internal server error."
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }